+++ lib/src/main/java/com/orange/ods/compose/component/appbar/top/OdsTopAppBar.kt
@@ -71,7 +71,7 @@ fun OdsTopAppBarInternal(
     title: String,
     modifier: Modifier = Modifier,
     navigationIcon: OdsTopAppBarNavigationIcon? = null,
-    actions: List<OdsComponentContent> = emptyList(),
+    actions: List<OdsComponentContent<*>> = emptyList(),
     overflowMenuActions: List<OdsTopAppBarOverflowMenuActionItem> = emptyList(),
     elevated: Boolean = true
 ) {