+++ changelog.md
@@ -5,6 +5,17 @@ All notable changes done in ODS library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.16.0](https://github.com/Orange-OpenSource/ods-android/compare/0.15.1...0.16.0) - 2023-10-10
+
+### Added
+
+- \[Lib\] Add `OdsModalDrawer` component documentation ([#652](https://github.com/Orange-OpenSource/ods-android/issues/652))
+
+### Changed
+
+- \[App\] Change buttons name ([#647](https://github.com/Orange-OpenSource/ods-android/issues/647))
+- \[Lib\] Update `OdsSlider` and `OdsSliderLockups` APIs ([#648](https://github.com/Orange-OpenSource/ods-android/issues/648))
+
 ## [0.15.1](https://github.com/Orange-OpenSource/ods-android/compare/0.15.0...0.15.1) - 2023-09-28
 
 ### Changed